Update a widget in shiny with an orientation

Pair of functions used in conjunction to update a imu_object widget

Usage

imu_proxy(id, session = shiny::getDefaultReactiveDomain())

imu_send_data(proxy, data)

Arguments

id

HTML element id that houses the html widget

session

shiny session - default to current session

proxy

imu_proxy output to be used as input to imu_send_data

data

a numeric unit 4-vector (quaternion) for the updated orientation